+++ app/controls/IdSpan.svelte
@@ -7,25 +7,36 @@
     export let selectable: boolean = false;
 
     $: suffix = id.rest.substring(0, 8 - id.prefix.length);
+    $: category = id.type == "CommitId" ? "commit" : id.is_divergent ? "divergent" : id.offset ? "hidden" : "change";
 </script>
 
 <span class="id" class:pronoun={pronoun || $currentTarget?.type == "Repository"} class:selectable>
-    <span class="prefix {id.type}">{id.prefix}</span>{suffix}
+    <span class="prefix {category}">{id.prefix}</span>{suffix}{#if id.type == "ChangeId" && id.offset}<span
+            class="suffix {category}">/{id.offset}</span
+        >{/if}
 </span>
 
 <style>
     .id {
         pointer-events: none;
-        color: var(--ctp-subtext1);
+        color: var(--ctp-subtext0);
         font-family: var(--stack-code);
     }
 
-    .ChangeId {
+    .commit {
+        color: var(--ctp-mauve);
+    }
+
+    .change {
         color: var(--ctp-pink);
     }
 
-    .CommitId {
-        color: var(--ctp-mauve);
+    .hidden {
+        color: var(--ctp-text);
+    }
+
+    .divergent {
+        color: var(--ctp-red);
     }
 
     .pronoun {
@@ -33,7 +44,8 @@
         pointer-events: none;
     }
 
-    .pronoun > .prefix {
+    .pronoun > .prefix,
+    .pronoun > .suffix {
         color: inherit;
         font-weight: bold;
     }